Исправил ожидаемый результат
